feat: 바로팟 상세 페이지 SEO 추가
- Server Component 수정, Client Component 추가
